Update docker.md
1. At present the existing start.ps1 fails to unregister the agent at the end of the run 
https://github.com/MicrosoftDocs/azure-devops-docs/issues/6019
This is due to the token being removed and then trying to be used. 
This replacement start.ps1 works and removes the agent.

2. There are other issues related to the file failing due to TLS1.2 not being added at the top. This should prevent or close those from reoccurrence. 

3. instructions for passing --once are very unclear and just adding --once to the docker run command fails and causes error.   


Please note, I am nowhere near an expert in docker, however these changes took me from not-working to working and I can see open issues for them still so figured I'd give a bash at helping out with the fix.  The issues have been open a while.